broke up raycasting.c into a number of separate files
