+++ Makefile
@@ -3,10 +3,10 @@
 #                                                         :::      ::::::::    #
 #    Makefile                                           :+:      :+:    :+:    #
 #                                                     +:+ +:+         +:+      #
-#    By: tparratt <tparratt@student.hive.fi>        +#+  +:+       +#+         #
+#    By: rboudwin <rboudwin@student.hive.fi>        +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/26 15:24:40 by tparratt          #+#    #+#              #
-#    Updated: 2024/09/10 11:08:57 by rboudwin         ###   ########.fr        #
+#    Updated: 2024/10/18 12:37:20 by rboudwin         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -23,7 +23,11 @@ SRC		=	srcs/main.c \
 			srcs/parsing4.c \
 			srcs/walls.c \
 			srcs/raycaster.c \
-			srcs/raycaster2.c
+			srcs/raycaster2.c \
+			srcs/setup_mlx.c \
+			srcs/init_images.c \
+			srcs/move_hooks.c \
+			srcs/draw_2d_map.c
 LIBFT		=	libft/libft.a
 
 MLX42_DIR   = ./MLX42
