+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: rboudwin <rboudwin@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/12 12:05:40 by tparratt          #+#    #+#             */
-/*   Updated: 2024/10/16 09:56:16 by rboudwin         ###   ########.fr       */
+/*   Updated: 2024/10/18 12:35:55 by rboudwin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -123,5 +123,11 @@ void	print_error(char *str, t_map *map);
 size_t	len_2d(char **arr);
 void	check_walls(t_map *map);
 void	setup_mlx(t_map *map);
+void	draw_2d_map(mlx_t *mlx, t_map *map, t_info *info);
+void	init_img(t_info *info);
+void	init_plyr(t_info *info, t_map *map);
+void	floor_and_ceiling(mlx_t *mlx, t_images *img, t_info *info, t_map *map);
+void	ft_movehook(void *param);
+void	draw_2d_player(mlx_t *mlx, t_map *map, t_info *info);
 
 #endif
